requirements.txt
Django==2.2.2
django-map-widgets==0.2.2
djangorestframework==3.9.4
Pillow==6.0.0
psycopg2-binary==2.8.3
pytz==2019.1
sorl-thumbnail==12.5.0
sqlparse==0.3.0
uWSGI==2.0.18
